Major refactor; CustomerBalance is now int rather than BigDecimal

+++ src/main/java/coinslot/ICoinSlot.java
@@ -1,12 +1,10 @@
 package main.java.coinslot;
 
-import java.math.BigDecimal;
-
 public interface ICoinSlot {
 	
 	public String getCoinName();
 
-	public BigDecimal getCoinValue();
+	public int getCoinValue();
 	
 	public void addCoin();
