+++ src/main/java/product/IProduct.java
@@ -1,10 +1,9 @@
 package main.java.product;
-import java.math.BigDecimal;
 
 public interface IProduct {
 	
 	public String getType();
 	
-	public BigDecimal getPrice();
+	public int getPrice();
 	
 }
